backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
comtypes==1.4.10
cryptography==45.0.4
distlib==0.3.9
Django==5.2.3
django-cors-headers==4.7.0
django-crispy-forms==2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
filelock==3.17.0
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
numpy==2.3.0
packaging==24.2
pillow==11.2.1
pipenv==2024.4.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
pypiwin32==223
pyttsx3==2.98
pywin32==308
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
setuptools==76.0.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.3.0
virtualenv==20.29.3
